+++ RELEASE_NOTES.rst
@@ -4,6 +4,7 @@ Next release
 All changes
 -----------
 
+- :pull:`366`: Add new westeros tutorial on modeling fossil resource supply curves.
 - :pull:`391`: Add a documentation page on :doc:`pre-requisite knowledge & skills <prereqs>`.
 - :pull:`389`: Fix a bug in :func:`.pyam.concat` using *non*-pyam objects.
 - :pull:`286`, :pull:`381`, :pull:`389`: Set ``duration_period`` in :meth:`.add_horizon`; add documentation of :doc:`time`.